CcdaParser: extract into its own file (Zeitwerk autoload)
CcdaParser lived as an inner class inside ccda_generator.rb, which is
a one-constant-per-file violation. Zeitwerk could only resolve
Lakeraven::EHR::CcdaParser if something loaded ccda_generator.rb first,
making CcdaImportsControllerTest seed-dependent: ~3/8 random seeds would
load the controller test before the generator test and raise
NameError: uninitialized constant Lakeraven::EHR::CcdaParser.

Extracted CcdaParser to app/services/lakeraven/ehr/ccda_parser.rb verbatim;
removed the inner class from ccda_generator.rb. Verified across 3 random
seeds plus the previously-flaky controller test in isolation.

Pinned the structural fix with a source_location-based regression test
so the parser can't drift back into another file without the test
catching it.
